# optimize topk on cpu using parallel and partial sort