@fmassa 2019-04-27
While not conflicting with this PR, https://github.com/pytorch/pytorch/pull/18344 is relevant, as it optimizes kthvalue for float, and one could use kthvalue to implement topk