@cpuhrsch 2019-04-28
@fmassa Have we look at [faiss](https://github.com/facebookresearch/faiss) as a source for the top-k implementation? In particular for GPU.