@mingfeima 2019-04-28
> While not conflicting with this PR, #18344 is relevant, as it optimizes kthvalue for float, and one could use kthvalue to implement topk

Yes, this one is not conflicting #18344. In case the `avx2` partition is faster than `std::partial_sort`, we can switch from `std::partial_sort` to `quick_select_template` very easily, just like the reference #19737