@fmassa 2019-04-28
@cpuhrsch I believe the kernels in faiss were (are?) the same as the ones in PyTorch. Indeed, @wickedfoo was the one who added support for topk on the GPU back in the Lua days, and I'm not sure if there are differences nowadays that would be good to port back