@soumith 2019-04-28
the ones in faiss are significantly different than pytorch, though jhj is the person who wrote both. The ones in faiss are also significantly more complex than what pytorch needs.