@cpuhrsch 2019-04-28
@mingfeima - could you also provide benchmarks for varying numbers of threads? In particular single core performance.