@mingfeima 2019-04-29
@cpuhrsch scaling result with `OMP_NUM_THREADS=1, 2, 4, 8, 16`.
before:
```
### topk: OMP_NUM_THREADS=1 ###
#[10, 10000] times: 0.749171 ms
#[10, 20000] times: 1.449934 ms
#[10, 40000] times: 2.854897 ms
#[10, 80000] times: 5.720388 ms
#[10, 160000] times: 11.393573 ms
#[10, 320000] times: 22.958385 ms
### topk: OMP_NUM_THREADS=2 ###
#[10, 10000] times: 0.748674 ms
#[10, 20000] times: 1.452205 ms
#[10, 40000] times: 2.847788 ms
#[10, 80000] times: 5.707714 ms
#[10, 160000] times: 11.470940 ms
#[10, 320000] times: 22.871153 ms
### topk: OMP_NUM_THREADS=4 ###
#[10, 10000] times: 0.746231 ms
#[10, 20000] times: 1.448718 ms
#[10, 40000] times: 2.868835 ms
#[10, 80000] times: 5.650503 ms
#[10, 160000] times: 11.451153 ms
#[10, 320000] times: 22.694164 ms
### topk: OMP_NUM_THREADS=8 ###
#[10, 10000] times: 0.745902 ms
#[10, 20000] times: 1.451694 ms
#[10, 40000] times: 2.853764 ms
#[10, 80000] times: 5.688034 ms
#[10, 160000] times: 11.397316 ms
#[10, 320000] times: 22.893178 ms
### topk: OMP_NUM_THREADS=16 ###
#[10, 10000] times: 0.743552 ms
#[10, 20000] times: 1.459475 ms
#[10, 40000] times: 2.864395 ms
#[10, 80000] times: 5.682968 ms
#[10, 160000] times: 11.432610 ms
#[10, 320000] times: 22.753903 ms
```

after:
```
### topk: OMP_NUM_THREADS=1 ###
#[10, 10000] times: 0.739171 ms
#[10, 20000] times: 1.347100 ms
#[10, 40000] times: 2.909180 ms
#[10, 80000] times: 5.716366 ms
#[10, 160000] times: 12.710830 ms
#[10, 320000] times: 28.739215 ms
### topk: OMP_NUM_THREADS=2 ###
#[10, 10000] times: 0.471362 ms
#[10, 20000] times: 0.796563 ms
#[10, 40000] times: 1.673725 ms
#[10, 80000] times: 3.335849 ms
#[10, 160000] times: 6.459378 ms
#[10, 320000] times: 15.167443 ms
### topk: OMP_NUM_THREADS=4 ###
#[10, 10000] times: 0.323723 ms
#[10, 20000] times: 0.518589 ms
#[10, 40000] times: 1.106971 ms
#[10, 80000] times: 2.216807 ms
#[10, 160000] times: 4.901155 ms
#[10, 320000] times: 11.800200 ms
### topk: OMP_NUM_THREADS=8 ###
#[10, 10000] times: 0.266110 ms
#[10, 20000] times: 0.396013 ms
#[10, 40000] times: 0.824147 ms
#[10, 80000] times: 1.694105 ms
#[10, 160000] times: 3.393351 ms
#[10, 320000] times: 8.041087 ms
### topk: OMP_NUM_THREADS=16 ###
#[10, 10000] times: 0.180306 ms
#[10, 20000] times: 0.267109 ms
#[10, 40000] times: 0.548724 ms
#[10, 80000] times: 1.096469 ms
#[10, 160000] times: 2.346745 ms
#[10, 320000] times: 5.103983 ms
```

This patch did parallelization only so won't help when using single core. But in case you use more than 2 cores, it helps.

In case we use parallelization, `partial_sort` performs slightly better than `quick select`, probably because `quick select` has an average time complexity of `o(n)` but the worst case is `o(n2)`. And in the thread pool you have wait for the every last one to finish. Anyway it should be very easy to switch the inner sorting algorithm in case we know a better option, like #18344