@wickedfoo 2019-04-29
This thread seems to be about the CPU but there was a question about the GPU.

The GPU top-k code in (Lua) Torch is different than the top-k code in Caffe2 is different than the top-k code in Faiss. I wrote all three.

The code in Torch supports arbitrary k (1 <= k <= n) via radix sort. It can be slow because it requires multiple passes over the input. The topk call in Torch needs to support arbitrary k, but it can have a specialization for small k (e.g., k <= 1024 or 2048).

The code in Caffe2 I think supports small k only, but could be wrong on that. It contains code from a very early version of Faiss.

The code in Faiss is fastest, but it only supports k <= 2048.