@VitalyFedyunin 2019-04-29
@mingfeima can you please run compare before, after for k=10,100,N/10,N/2, N-5