@mingfeima 2019-04-30
> @mingfeima can you please run compare before, after for k=10,100,C/10,C/2, C-5

@VitalyFedyunin 
1. Any range for C?
2. Number of threads? (This patch wont help for single thread)
3. C-5 refers to sorting the whole channel except the last five?
4. In case k is a relatively large number (for example C/2), `torch.(sorted=True)` or `torch.(sorted=False)`?

I was assuming that k is not a very large number (in beam search it is `beam_size * 2` or even less), so whether sorting the k values or not won't make any difference. But in case you have scenarios of dealing with relatively large k, it will.