@fmassa 2019-04-30
@mingfeima it would also be good to run it with a single thread, to assess the impact of performing a copy of the data into a vector to feed to the `partial_sort`.

Here is [another example](https://github.com/facebookresearch/maskrcnn-benchmark/blob/eb4d3352be7e968b96260c8999a331e8431da95f/maskrcnn_benchmark/modeling/rpn/inference.py#L95) where `topk` is used (for mask-rcnn).
The typical sizes  for the tensor `objectness` are:
```
objectness = torch.rand(2, 200000)
pre_nms_top_n = 6000
objectness, topk_idx = objectness.topk(pre_nms_top_n, dim=1, sorted=True)
```