@VitalyFedyunin 2019-05-01
1. Try to pick various C. 10,000  40,000  320,000
2. Compare single thread and 20 threads.
3. Yes
4. sorted=False