@mingfeima 2019-05-23
@VitalyFedyunin Hi, all the proposed patterns have been tested to have a more generic converage:
1. C = 10000, 40000, 320000
2. K = 10, 50, 100, C/10, C/2, C-5
3. Test with 20 threads and 1 thread
4. Test with Sorted=True and Sorted=False

All the details including:
1. test pattern
2. benchmark scripts
3. performance results
4. optimization strategy
5. topk design from other frameworks

are listed in the [gist](https://gist.github.com/mingfeima/115e88f9196447db2d91f8772c924f31), otherwise it's too much to show in this thread.


To sum up a little bit, with the latest code in this pr, `topk()` compared to original:
1. 20 threads and sorted=True: **8.8x** speedup.
2. 20 threads and sorted=False: **7.5x** speedup.
3. 1 thread and sorted=True: **1.8x** speedup.
4. 1 thread and sorted=False: **1.7x** speedup.

Speedup from single thread runs come from `std::partial_sort` over quick select.