@cpuhrsch 2019-05-31
cc @VitalyFedyunin for TensorIterator.

@mingfeima - could I ask you to rebase the PR so we can import it?