@VitalyFedyunin 2019-05-31
Writing `dim_apply` implementation for TensorIterator is nice thing, but might be out of scope for this PR. I will add it in our _wishlist table_