@mingfeima 2019-06-26
@VitalyFedyunin The CPU kernels have been moved to `native/cpu/SortingKernel.cpp`

Perf is slightly better compared with last version,

1. 20 threads and sorted=True: **9.4x** speedup (last version **8.8x**).
2. 20 threads and sorted=False: **8.2x** speedup (last version **7.5x**).
3. 1 thread and sorted=True: **1.8x** speedup (last version **1.8x**).
4. 1 thread and sorted=False: **1.7x** speedup (last version **1.7x**).